Problem in testing
I test in fresh laravel 5.7 and try to run tests
But I have problem with that TestCase inside package is not working at all.
i change it and problem solved .
I make test fil easier if i'm worng tell me about it .
