--- tests/TestCase.php
@@ -1,35 +1,0 @@
-<?php
-
-namespace Tests;
-
-use Orchestra\Testbench\TestCase as BaseTestCase;
-
-abstract class TestCase extends BaseTestCase
-{
-    public function setUp()
-    {
-        parent::setUp();
-
-        $this->loadMigrationsFrom(__DIR__ . '/database/migrations');
-
-        $this->artisan('migrate');
-    }
-
-    protected function getEnvironmentSetUp($app)
-    {
-        $app['config']->set('database.default', 'testbench');
-
-        $app['config']->set('database.connections.testbench', [
-            'driver'   => 'sqlite',
-            'database' => ':memory:',
-            'prefix'   => '',
-        ]);
-    }
-
-    protected function getPackageProviders($app)
-    {
-        return [
-            \Watson\Aggregate\AggregateServiceProvider::class,
-        ];
-    }
-}
